lab02 layout, header, sidemenu.
