+++ json-server/db.json
@@ -12,7 +12,7 @@
       "title": "Introduction to Styling with CSS3",
       "url": "https://file-examples.com/wp-content/uploads/2017/04/file_example_MP4_1280_10MG.mp4",
       "duration": "0:18",
-      "approved": true
+      "approved": false
     },
     {
       "id": 2,
