+++ lab02_angular8/smart-player/src/app/app-routing.module.ts
@@ -4,9 +4,10 @@ import { HomeComponent } from './home/home.component';
 import { CourseComponent } from './course/course.component';
 
 const routes: Routes = [
-  { path: '', component: HomeComponent, pathMatch: 'full' },
+  { path: '', redirectTo: 'course', pathMatch: 'full' },
+  { path: 'watch', component: HomeComponent, pathMatch: 'full' },
   { path: 'course', component: CourseComponent, pathMatch: 'full' },
-  { path: '**', component: HomeComponent },
+  { path: '**', component: CourseComponent },
 ];
 
 @NgModule({
