+++ lab02_angular8/smart-player/src/app/app.component.html
@@ -1,1 +1,7 @@
-<router-outlet></router-outlet>
+<app-layout-header></app-layout-header>
+<div class="body">
+  <app-sidemenu></app-sidemenu>
+  <main>
+    <router-outlet></router-outlet>
+  </main>
+</div>
