+++ lab02_angular8/smart-player/src/app/core/core.module.ts
@@ -1,8 +1,23 @@
 import { NgModule } from '@angular/core';
 import { SharedModule } from '../shared/shared.module';
+import { CourseService, ReqInterceptor, ResInterceptor } from '.';
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
 
 @NgModule({
   imports: [SharedModule],
   declarations: [],
+  providers: [
+    CourseService,
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: ReqInterceptor,
+      multi: true,
+    },
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: ResInterceptor,
+      multi: true,
+    },
+  ],
 })
 export class CoreModule {}
