+++ lab02_angular8/smart-player/src/app/core/index.ts
@@ -1,1 +1,2 @@
 export * from './interceptors';
+export * from './services';
