+++ lab02_angular8/smart-player/src/app/core/models/course.model.ts
@@ -1,0 +1,7 @@
+export interface Course {
+  id: number;
+  title: string;
+  url: string;
+  duration: string;
+  approved: boolean;
+}
