+++ lab02_angular8/smart-player/src/app/core/models/index.ts
@@ -1,0 +1,1 @@
+export * from './course.model';
