+++ lab02_angular8/smart-player/src/app/core/services/index.ts
