+++ lab02_angular8/smart-player/src/app/home/home.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 
 import { HomeRoutingModule } from './home-routing.module';
-import { SharedModule } from '../shared/shared.module';
+import { SharedModule } from '../shared';
 import {
   VideoPlayerComponent,
   PlayerComponent,
