+++ lab02_angular8/smart-player/src/app/home/video-player/controls/controls.component.css
@@ -1,7 +1,7 @@
 #videoControls {
   position: absolute;
   min-width: 375px;
-  bottom: 6px;
+  bottom: 0;
   width: 100%;
 }
