+++ lab02_angular8/smart-player/src/app/home/video-player/player/player.component.css
@@ -1,5 +1,6 @@
  video {
    min-width: 375px;
    width: 100%;
-   height: auto;
+   height: 100%;
+   object-fit: fill;
  }
