+++ lab02_angular8/smart-player/src/app/home/video-player/playlist/playlist.component.html
@@ -1,10 +1,20 @@
 <div class="section-title">
-  Full stack courses
+  Chapter 01 Introduction
 </div>
 <ul class="course-list">
-  <li class="course" *ngFor="let c of courses; let i = index; trackBy: c ? c.id : null">
-    <a class="link" [class.active]="selectedIndex === i" href="javascript:void(0)" (click)="handleSelection(i)">
-      <div class="flex items-center"><i class="fab fa-youtube"></i> {{ c.title }}</div>
+  <li
+    class="course"
+    *ngFor="let c of courses; let i = index; trackBy: c ? c.id : null"
+  >
+    <a
+      class="link"
+      [class.active]="selectedIndex === i"
+      href="javascript:void(0)"
+      (click)="handleSelection(i)"
+    >
+      <div class="flex items-center">
+        <i class="fab fa-youtube"></i> {{ c.title }}
+      </div>
       <div>
         <span class="time">{{ c.duration }}</span>
       </div>
