+++ lab02_angular8/smart-player/src/app/home/video-player/video-player.component.css
@@ -2,4 +2,5 @@
   /* background: #000; */
   position: relative;
   z-index: 2147483647;
+  background-color: #000;
 }
