+++ lab02_angular8/smart-player/src/app/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './layout';
+export * from './shared.module';
