+++ lab02_angular8/smart-player/src/app/shared/layout/header.component.html
@@ -1,0 +1,10 @@
+<nav class="navbar navbar-light">
+  <a class="navbar-brand">
+    <span class="mr-2">FSD Course Center</span>
+    <i class="fas fa-bars"></i>
+  </a>
+  <div>
+    <i class="far fa-user"></i>
+    uniquejava
+  </div>
+</nav>
