+++ lab02_angular8/smart-player/src/app/shared/layout/header.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-layout-header',
+  templateUrl: './header.component.html',
+  styleUrls: ['./header.component.scss'],
+})
+export class HeaderComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit() {}
+}
