+++ lab02_angular8/smart-player/src/app/shared/layout/index.ts
@@ -1,0 +1,2 @@
+export * from './header.component';
+export * from './sidemenu.component';
