+++ lab02_angular8/smart-player/src/app/shared/layout/sidemenu.component.scss
@@ -1,0 +1,33 @@
+nav {
+  height: 100%;
+  min-height: calc(100vh - 56px);
+}
+
+.menu {
+  margin: 0;
+  padding: 0;
+  width: 180px;
+  height: 100%;
+  list-style-type: none;
+  border-right: solid 1px #e6e6e6;
+}
+
+.menu-item {
+  padding: 0 20px;
+  height: 56px;
+  line-height: 56px;
+  font-size: 14px;
+  color: #303133;
+  cursor: pointer;
+  &:hover {
+    outline: 0;
+    background-color: rgb(125, 197, 156, 0.2);
+  }
+  &.active {
+    outline: 0;
+    background-color: rgb(125, 197, 156, 0.1);
+  }
+  i {
+    margin-right: 0.5rem;
+  }
+}
