+++ lab02_angular8/smart-player/src/index.html
@@ -9,12 +9,12 @@
 
     <link rel="icon" type="image/x-icon" href="favicon.ico" />
 
-    <link
+    <!-- <link
       rel="stylesheet"
       href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
       integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T"
       crossorigin="anonymous"
-    />
+    /> -->
     <link
       rel="stylesheet"
       href="https://use.fontawesome.com/releases/v5.9.0/css/all.css"
@@ -23,6 +23,6 @@
     />
   </head>
   <body>
-    <app-root></app-root>
+    <app-root> Loading... </app-root>
   </body>
 </html>
